Clean up mysql reconnect.
In general reconnect is dangerous when mixed with transactions. (i.e. during
a transaction if the connection is lost and a reconnect occurs then the
previous portion of the transaction is rolled back on the server and that
portion of the transaction is lost without the upper levels of software
knowing that it has occured, very bad) Obviously when transactions are
disabled (e.g. autocommit True) then this is not applicable and reconnect
is quite useful.

- By default disable reconnect, so it only gets used if someone knows what
  they're getting into.

- Add support so that reconnect can be enabled for a certain number of
  commands, therefore allowing reconnect on only the first command in
  a transaction.

- Add a connection check which drives reconnect and only occurs before the
  SQL command is sent to the server. (e.g. do not reconnect on a connection
  error during a command, otherwise the command is reissued and you can have
  a cascade of identical long running commands)

- The connection check uses a non-blocking poll on the socket when no data
  readiness is expected (before the command is sent) to see if an error is
  present. The data is buffered if an edge case exists where data should be
  present it will be preserved until the appropriate code path is taken.