beginnings of the tree.pull_prefix() implementation including node.split()
currently this leaves a vertical trail of death and destruction behind it,
but we won't need to clean up the left side -- we already return an iterator
over the left side and that's all we need. still to implement is bringing
the cut edge of the right side (now the whole tree) back into balance.